First stage of app

+++ Training018/src/main/java/pl/edu/bogdan/training/unittests/Board.java
@@ -3,7 +3,7 @@ package pl.edu.bogdan.training.unittests;
 public interface Board {
 	double getWidth();
 	double getHeight();
-	double getXOriginCoordinate();
-	double getYOriginCoordinate();
+	int getXOriginCoordinate();
+	int getYOriginCoordinate();
 	boolean containsPoint(double x, double y);
 }
